Reduces the number of viewports pa11y tests against
Just a test to find out if the number of tests we are trying to run is
causing failures.

Stream pages were we test a lot of urls is struggling to pass CI builds
and I am wondering if this will help at all. If it does we would need to
have a think about how else we might improve this,